backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Marshmallow==3.21.2
python-dotenv==1.0.1
pg8000==1.31.2
pytz==2024.1
firebase-admin==6.5.0
APScheduler==3.10.4
requests-oauthlib==2.0.0
google-generativeai>=0.7.2
tensorflow==2.20.0
tensorflow-hub==0.16.1
